backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx==0.28.1
redis==5.2.1
sqlalchemy==2.0.36
pymysql==1.1.1
requests==2.32.3
python-dotenv==1.0.1
